feat: add auth with jwt
